feat: move selection from class Board
Since selection was a part for the player process, the method for selecting
the slot of the board was move to the class Player. The method was adjusted
so that it will only accept digits and adding conditions for checking if
the slot is still available or not.

Further optimization is needed and if there is a need to move the process
through which the slots is still available or not as each player object
contains different states.