feat(!): add loop to play the game
The game is now playable with players taking turns in choosing a board
slot. Each board slot chosen by the player will be added to their array of
player_choices. This choices were continuously compared to a different array.

If a player has a combination which is included in a different array, the
player wins that game.