fix: sort the player choices before combination
Combination method of the array class generates all possible outcomes with
the given length, the problem is that it generates the array based on the
order of the elements inside the array. So an array starting at 5 will have
its array combination also starts at 5. This causes an error as the
arrangement of the lines array is already fix and was sorted from lowest
to highest.

This was fixed by first sorting the array then doing the enumerable
combination.